most. convoluted. solution. ever.  Now, the sell button only appears on your squad's players
to implement this, I had to add another content script that only ran in
the top frame.  Its sole purpose is to send the contents of the
innerHTML in the top frame to the background page, which stores it.
Then, the other content script inside the frame sends a request for the
innerHTML previously stored and continues on the response.  Because all
of this is asynchronous, I also had to account for the chance that the
request for the innerHTML happened before it had been passed, and store
the response function just in case.  but it works!